Move unit test to cdc-core
