+++ shardingsphere-kernel/shardingsphere-schedule/shardingsphere-cdc/shardingsphere-cdc-core/src/main/java/org/apache/shardingsphere/cdc/core/util/ReflectionUtil.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.scaling.core.util;
+package org.apache.shardingsphere.cdc.core.util;
 
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
